minor bugs fixed
