+++ realize/cli_test.go
@@ -1,12 +1,12 @@
 package realize
 
 import (
+	"bytes"
+	"log"
 	"os"
+	"strings"
 	"testing"
 	"time"
-	"strings"
-	"log"
-	"bytes"
 )
 
 func TestRealize_Stop(t *testing.T) {
@@ -21,7 +21,7 @@ func TestRealize_Stop(t *testing.T) {
 
 func TestRealize_Start(t *testing.T) {
 	r := Realize{}
-	go func(){
+	go func() {
 		time.Sleep(100)
 		close(r.exit)
 		_, ok := <-r.exit
@@ -36,7 +36,7 @@ func TestRealize_Prefix(t *testing.T) {
 	r := Realize{}
 	input := "test"
 	result := r.Prefix(input)
-	if len(result) <= 0 && !strings.Contains(result,input){
+	if len(result) <= 0 && !strings.Contains(result, input) {
 		t.Error("Unexpected error")
 	}
 }
@@ -47,7 +47,7 @@ func TestSettings_Read(t *testing.T) {
 	w := LogWriter{}
 	input := ""
 	int, err := w.Write([]byte(input))
-	if err != nil || int > 0{
-		t.Error("Unexpected error", err, "string lenght should be 0 instead",int)
+	if err != nil || int > 0 {
+		t.Error("Unexpected error", err, "string lenght should be 0 instead", int)
 	}
 }
